port other patient details enums
